fix(shared,testing): remove @clerk/types dependency from shared package
