+++ libs/shared/package.json
@@ -44,7 +44,6 @@
     "class-validator": "0.14.0"
   },
   "devDependencies": {
-    "@clerk/types": "^4.6.1",
     "@types/bluebird": "^3.5.24",
     "@types/jest": "29.5.2",
     "jest": "^27.1.0",
