+++ libs/shared/src/types/auth/auth.types.ts
@@ -1,4 +1,3 @@
-import { JwtPayload } from '@clerk/types';
 import { SignUpOriginEnum } from '../analytics';
 
 export interface IJwtClaims {
@@ -13,17 +12,6 @@ export interface IJwtClaims {
   iss?: string;
   scheme: ApiAuthSchemeEnum.BEARER | ApiAuthSchemeEnum.API_KEY;
 }
-
-// JWT payload + custom claims
-export type ClerkJwtPayload = JwtPayload & {
-  _id: string;
-  email: string;
-  lastName: string;
-  firstName: string;
-  profilePicture: string;
-  externalId?: string;
-  externalOrgId?: string;
-};
 
 export type UserSessionData = IJwtClaims & { environmentId: string };
